[NN-0037] Improve GUI

+++ NeuralNetwork/src/Header Files/neuralNetwork.h
@@ -58,7 +58,7 @@ class NeuralNetwork
         void train(const std::vector<float> &inputs, const std::vector<float> &desired);
         std::vector<float> output(const std::vector<float>& inputs);
 
-        void calculateClusteringRateForRegressionProblem(const std::vector<float> &inputs, const std::vector<int> &desired);
+        void calculateClusteringRateForRegressionProblem(const std::vector<float>& inputs, const std::vector<float>& desired);
         void calculateClusteringRateForClassificationProblem(const std::vector<float> &inputs, const uint classNumber);
 
         void addANeuron(uint layerNumber);
